refactor: create hook for characters page
