+++ src/pages/characters/characters-page.tsx
@@ -1,28 +1,18 @@
-import React, { useEffect } from "react";
+import React from "react";
 import { CharactersList, SearchBar } from "../../components";
-import { useFetchCharacters } from "./hooks/use-fetch-characters";
-import { useSearchBar } from "./hooks/use-search-bar";
+
 import "./characters-page.css";
-import { useFavoriteCharacters } from "./hooks/use-favorite-characters";
+import { useCharactersPage } from "./hooks/use-characters-page";
 
 const CharactersPage = () => {
-  const { characters, fetchCharacters } = useFetchCharacters();
-  const { isSearching, results, searchCriteria, onChangeSearchCriteria } =
-    useSearchBar();
-  const { isFavorite, toggleLike: toggleLikeCharacter } =
-    useFavoriteCharacters();
-
-  useEffect(() => {
-    fetchCharacters();
-    //  eslint-disable-next-line
-  }, []);
-
-  const charactersList = (searchCriteria ? results : characters).map(
-    (character) => ({
-      ...character,
-      liked: isFavorite(character.id)
-    })
-  );
+  const {
+    isSearching,
+    searchCriteria,
+    charactersList,
+    toggleLike,
+    showCharacterDetails,
+    onChangeSearchCriteria
+  } = useCharactersPage();
 
   return (
     <main className="characters-page">
@@ -34,8 +24,8 @@ const CharactersPage = () => {
       />
       <CharactersList
         characters={charactersList}
-        onCharacterClick={(id) => console.log("Character->click", id)}
-        onCharacterLike={toggleLikeCharacter}
+        onCharacterClick={showCharacterDetails}
+        onCharacterLike={toggleLike}
       />
     </main>
   );
